\name{epprobit-package}
\alias{epprobit-package}
\alias{epprobit}
\docType{package}
\title{
\packageTitle{epprobit}
}
\description{
\packageDescription{epprobit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{epprobit}
\packageIndices{epprobit}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{epprobit}

Maintainer: \packageMaintainer{epprobit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
